Basic IIO support for the ZaiNar DMA loopback driver
This is the first rough basic implementation which adds IIO to the
AXI dma test code.  The additions are directly from the in kernel
examples contained in drivers/iio/dummy/ they have been copied and added
to our code with the full intention to update functionality for our
purposes.

The iio_dummy is being used to understand the frame work and as a
starting point for future development.

Renamed the kernel example DMA test loopback test to zainar_axidmatest
as we will be using it as our baseline/starting point for the driver
going forward.